// src/lib/sanity/types.ts
export type GetWorkoutsQueryResult = Array<{
  _id: string;
  date: string | null;
  duration: number | null;
  exercises: Array<{
    exercise: {
      _id: string;
      name: string | null;
    } | null;
    sets: Array<{
      reps: number | null;
      weight: number | null;
      weightUnit: "kg" | "lbs" | null;
      _type: "set";
      _key: string;
    }> | null;
    _type: "exerciseEntry";
    _key: string;
  }> | null;
}>;